adding repository interfaces and services

+++ src/main/java/tn/esprit/nejd_bedoui_project_4twin7/Models/Foyer.java
@@ -14,7 +14,7 @@ public class Foyer implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idFoyer;
+    private Long idFoyer;
     private String nomFoyer;
     private Long capaciteFoyer;
